+++ src/main/java/net/oneandone/pommes/search/CentralSearch.java
@@ -22,7 +22,7 @@ import net.oneandone.maven.summon.api.Maven;
 import net.oneandone.pommes.database.PommesQuery;
 import net.oneandone.pommes.database.Project;
 import net.oneandone.pommes.descriptor.MavenDescriptor;
-import net.oneandone.pommes.scm.Scm;
+import net.oneandone.pommes.scm.ScmUrl;
 import net.oneandone.sushi.fs.World;
 import net.oneandone.sushi.fs.http.HttpNode;
 import org.apache.maven.project.MavenProject;
@@ -99,7 +99,7 @@ public class CentralSearch {
 
     private Project projectOpt(Artifact artifact) throws IOException {
         MavenProject p;
-        String scm;
+        ScmUrl scm;
 
         try {
             p = maven.loadPom(artifact);
@@ -111,7 +111,7 @@ public class CentralSearch {
             System.out.println("TODO: no scm: " + artifact);
             return null;
         }
-        return MavenDescriptor.mavenToPommesProject(p, "central", p.getFile().getPath(), p.getVersion(), Scm.createUrl(scm));
+        return MavenDescriptor.mavenToPommesProject(p, "central", p.getFile().getPath(), p.getVersion(), scm);
     }
 
     private static String getString(JsonObject obj, String member) throws IOException {